Version 1.1.1
Creates of Updates a shapefile from the Marine Debris Tracker
Application from the University of Georgia see
(http://www.marinedebris.engr.uga.edu/). Create a user account, and
download the most recent csv with all of the  data. Use that csv as an
input for this program. If the marine debris shapefile has not been
created, the script will create a new shapefile from the csv; if the
file has been created, it will update the shapefile with any of the
entries which have not been added to the shapefile yet.